moved everything to a simple directory
