+++ src/buffered/query.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, error::Error, fmt::Display};
+use std::{error::Error, fmt::Display};
 
 use tokio::sync::{
     mpsc,
